+++ lib/torque_api/objects/return_rma_response.rb
@@ -1,0 +1,5 @@
+module TorqueAPI
+  module Objects
+    class ReturnRmaResponse < Base; end
+  end
+end