+++ lib/torque_api/resources/return_rma_resource.rb
@@ -1,8 +1,8 @@
 module TorqueAPI
   class ReturnRmaResource < Resource
     def list
-      data = get_request("returnRma")
-      Array(data).map { |item| Objects::ReturnRma.new(item) }
+      data = get_request("returnRma/100")
+      Objects::ReturnRmaResponse.new(data)
     end
   end
 end